Agregado update y delete de usuarios, superadmin

+++ public/super-admin/super-admin.component.js
@@ -18,9 +18,11 @@ angular.
 				self.addUser = function(){
 					self.data = {
 						matricula: self.username,
+						nombre: self.name,
+						//apellido1: self.apellido1,
+						//apellido1: self.apellido2,
 						password: self.password,
-						password2: self.password2,
-						nombre: self.name
+						password2: self.password
 					}
 
 					$http({
@@ -28,6 +30,7 @@ angular.
 						url: 'http://localhost:8000/api/v1.0/users',
 						data: self.data
 					}).success(function(data){
+						alert("Usuario agregado exitosamente!");
 						window.location.reload();
 					}).error(function(){
 						alert('Ocurrio un error');
@@ -35,15 +38,46 @@ angular.
 				}
 
 				self.deleteUser = function(username){
-					alert("BOrrar a " + username);
+					var r = confirm("¿Seguro que deseas borrar al usuario "+ username + "?");
+					if( r == true){
+						$http({
+							method: 'DELETE',
+							url: 'http://localhost:8000/api/v1.0/users/' + username,
+						}).success(function(data){
+							alert("Usuario eliminado exitosamente!");
+							window.location.reload();
+						}).error(function(){
+							alert('Ocurrio un error al eliminar');
+						});
+					}
+				}
+				self.updateUser = function(user){
+					console.log(user);
+					self.data = {
+						//matricula: user.matricula,
+						nombre: user.nombre,
+						apellido1: user.apellido1,
+						apellido2: user.apellido2,
+						password: user.password,
+						password2: user.password
+					}
+					console.log("data");
+					console.log(self.data);
+
 					$http({
-						method: 'DELETE',
-						url: 'http://localhost:8000/api/v1.0/users/' + username,
-					}).success(function(data){
+						method: 'PUT',
+						url: 'http://localhost:8000/api/v1.0/users/' + user.matricula,
+						data: self.data
+					}).success(function(user){
+						alert("Usuario actualizado exitosamente");
 						window.location.reload();
 					}).error(function(){
-						alert('Ocurrio un error al eliminar');
+						alert("Error, no se puede actualizar el usuario");
 					});
+				}
+
+				self.refreshPage = function(){					
+					window.location.reload();
 				}
 
 			}]
